Set and restore raw mode using x/term on Linux
On Linux, set the terminal to raw mode and restore it using MakeRaw()
and Restore() in x/term respectively, which is used upstream in place of
x/sys/unix.IoctlSetTermios() or syscall.Syscall6() in other platforms.

MakeRaw() performs the same changes with certain termios fields, but
copies the whole struct prior instead of those fields only.